issrr/requirements.txt
Pillow>=10.0.0
pyzmq>=25.0.0
requests>=2.31.0
